+++ pkg/azuremachine/common_test.go
@@ -1,0 +1,54 @@
+package azuremachine
+
+import (
+	"encoding/json"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	providerv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
+)
+
+func azureMachineRawObject(sshKey string) []byte {
+	mp := providerv1alpha3.AzureMachine{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "AzureMachine",
+			APIVersion: "infrastructure.cluster.x-k8s.io/v1alpha3",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "ab123",
+			Namespace: "default",
+			Labels: map[string]string{
+				"azure-operator.giantswarm.io/version": "5.0.0",
+				"giantswarm.io/cluster":                "ab123",
+				"cluster.x-k8s.io/cluster-name":        "ab123",
+				"cluster.x-k8s.io/control-plane":       "true",
+				"giantswarm.io/machine-pool":           "ab123",
+				"giantswarm.io/organization":           "giantswarm",
+				"release.giantswarm.io/version":        "13.0.0",
+			},
+		},
+		Spec: providerv1alpha3.AzureMachineSpec{
+			AvailabilityZone: providerv1alpha3.AvailabilityZone{},
+			Image: &providerv1alpha3.Image{
+				Marketplace: &providerv1alpha3.AzureMarketplaceImage{
+					Publisher:       "kinvolk",
+					Offer:           "flatcar-container-linux-free",
+					SKU:             "stable",
+					Version:         "2345.3.1",
+					ThirdPartyImage: false,
+				},
+			},
+			Location: "westeurope",
+			OSDisk: providerv1alpha3.OSDisk{
+				OSType:     "Linux",
+				DiskSizeGB: 50,
+				ManagedDisk: providerv1alpha3.ManagedDisk{
+					StorageAccountType: "Premium_LRS",
+				},
+			},
+			SSHPublicKey: sshKey,
+			VMSize:       "Standard_D4s_v3",
+		},
+	}
+	byt, _ := json.Marshal(mp)
+	return byt
+}