+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Check that SSH keys are not set in AzureMachine and AzureMachinePool CRs.
 - Add mutating webhook to set storage account type in the AzureMachinePool CR if unset.
 
 ### Changed